compio: isolate unsafe wrappers
Move single-thread cell access, local stream sharing, and raw syscall
calls behind small wrapper modules. Add focused unit tests for the pure
cell wrappers so they can be run under Miri.
